+++ compiler/testData/loadJava/compiledKotlinWithStdlib/annotations/JvmFieldInInterfaceCompanion.kt
@@ -1,0 +1,11 @@
+package test
+
+interface I {
+    companion object {
+        @JvmField
+        val x = "x"
+
+        @JvmField
+        val y = "y"
+    }
+}